Splits template.js -> util/content + new “content” gulp task
