+++ package.json
@@ -50,6 +50,7 @@
     "glob": "^4.3.5",
     "gulp": "^3.8.10",
     "gulp-changed": "^1.1.0",
+    "gulp-clone": "^1.0.0",
     "gulp-compile-handlebars": "^0.4.4",
     "gulp-connect": "^2.2.0",
     "gulp-filesize": "0.0.6",
@@ -65,9 +66,11 @@
     "html-template": "^1.2.1",
     "lodash": "^2.4.1",
     "markdown-it": "^3.0.4",
+    "merge-stream": "^0.1.7",
     "moment-timezone": "^0.3.0",
     "require-dir": "^0.1.0",
     "run-sequence": "^1.0.2",
+    "through2": "^0.6.3",
     "vinyl-buffer": "^1.0.0",
     "vinyl-map": "^1.0.1",
     "vinyl-source-stream": "^1.0.0"
